minor fixes

+++ utils/calculator_tools.py
@@ -19,7 +19,7 @@ def multiply(a: int, b: int) -> int:
 
 
 @tool 
-def divide(a: int, b: int) -> int:
+def divide(a: int, b: int) -> float:
     '''Divide a (int) by b (int).'''
     return a/b
